analysis/revenue-forecast/requirements.txt
numpy>=1.26
pandas>=2.1
statsmodels>=0.14
matplotlib>=3.8
pyyaml>=6.0
